+++ src/test/java/nextstep/member/ThemeE2ETest.java
@@ -1,6 +1,7 @@
 package nextstep.member;
 
 import io.restassured.RestAssured;
+import nextstep.member.dto.MemberRequestDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
